refactor: spread store tests into multiple files
