+++ crates/store/tests/helpers.rs
@@ -1,3 +1,4 @@
+use domain::ParamValue;
 use sqlx::sqlite::SqlitePoolOptions;
 use store::Store;
 
@@ -8,4 +9,8 @@ pub async fn in_memory_store() -> Store {
         .await
         .unwrap();
     Store::from_pool(pool).await.unwrap()
+}
+
+pub fn make_params() -> std::collections::HashMap<String, ParamValue> {
+    [("speed".to_string(), ParamValue::Number(2.0))].into()
 }
